requirements.txt
crew-py
python-dotenv
litellm
pyyaml 